+++ DesktopEdge/Views/ItemRenderers/MenuEditSearch.xaml
@@ -1,0 +1,31 @@
+﻿<UserControl x:Class="ZitiDesktopEdge.MenuEditSearch"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ZitiDesktopEdge"
+             mc:Ignorable="d" 
+             MaxWidth="360"
+             d:DesignHeight="31" 
+             Height="31" 
+             HorizontalAlignment="Stretch" 
+             d:DesignWidth="350" 
+             Cursor="Hand">
+    <StackPanel Orientation="Vertical" Height="31">
+        <Grid Height="30">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="1*"/>
+            </Grid.ColumnDefinitions>
+            <TextBox Name="MainLabel" Background="White" Height="30" FontWeight="Bold" IsReadOnly="True" HorizontalAlignment="Stretch" BorderThickness="0" ClipToBounds="True" Grid.Column="0" VerticalAlignment="Center" Padding="0,6,3,0" FontSize="12" FontFamily="pack://application:,,,/Assets/Fonts/#Open Sans" Foreground="Black"></TextBox>
+            <TextBox Name="MainEdit" Background="White" HorizontalAlignment="Stretch" Height="30" Grid.Column="1" BorderThickness="1" TextAlignment="Left" Margin="5,0,0,0" Padding="5,3,3,5" FontSize="12" FontFamily="pack://application:,,,/Assets/Fonts/#Open Sans" VerticalContentAlignment="Center" SpellCheck.IsEnabled="True" MaxLength="100" KeyUp="MainEdit_KeyUp">
+                <TextBox.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="10"/>
+                    </Style>
+                </TextBox.Resources>
+            </TextBox>
+            <Label x:Name="ClearButton" Grid.Column="1" FontFamily="pack://application:,,,/Assets/Fonts/#Open Sans" Visibility="Collapsed" Margin="0,2,2,0" Foreground="LightGray" Content="clear" HorizontalAlignment="Right" Cursor="Hand" MouseUp="Label_MouseUp"></Label>
+        </Grid>
+    </StackPanel>
+</UserControl>
