+++ DesktopEdge/Views/ItemRenderers/MenuEditToggle.xaml
@@ -1,0 +1,23 @@
+﻿<UserControl x:Class="ZitiDesktopEdge.MenuEditToggle"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ZitiDesktopEdge"
+             mc:Ignorable="d" MaxWidth="360"
+             d:DesignHeight="31" 
+             Height="31"
+             HorizontalAlignment="Stretch" 
+             d:DesignWidth="350">
+    <StackPanel Orientation="Vertical" Height="31">
+        <Grid Height="30">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="1*"/>
+                <ColumnDefinition Width="40"/>
+            </Grid.ColumnDefinitions>
+            <TextBox Name="MainLabel" Background="White" Height="30" FontWeight="Bold" IsReadOnly="True" HorizontalAlignment="Stretch" BorderThickness="0" ClipToBounds="True" Grid.Column="0" VerticalAlignment="Center" Padding="0,6,3,0" FontSize="12" FontFamily="pack://application:,,,/Assets/Fonts/#Open Sans" Foreground="Black"></TextBox>
+            <local:Toggler x:Name="ToggleField" Grid.Column="1" HorizontalAlignment="Right" OnToggled="Toggled" Margin="0,0,5,0"></local:Toggler>
+        </Grid>
+    </StackPanel>
+</UserControl>
+
