+++ DesktopEdge/Views/ItemRenderers/ServiceInfo.xaml.cs
@@ -12,6 +12,8 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using ZitiDesktopEdge.Models;
+using ZitiDesktopEdge.ServiceClient;
 
 namespace ZitiDesktopEdge {
     /// <summary>
@@ -19,49 +21,29 @@ namespace ZitiDesktopEdge {
     /// </summary>
     public partial class ServiceInfo: UserControl {
 
-		private string _label = "";
-		private string _warning = "";
+		private ZitiService _info;
 
-		public string Warning {
+		public delegate void Mesage(string message);
+		public event Mesage OnMessage;
+		public delegate void Details(ZitiService info);
+		public event Details OnDetails;
+
+		public ZitiService Info { 
 			get {
-				return _warning;
+				return _info;
 			}
 			set {
-				_warning = value;
-				if (_warning.Length>0) {
-					WarnIcon.ToolTip = _warning;
+				this._info = value;
+				MainEdit.ToolTip = this._info.Url;
+				MainEdit.Text = this._info.Url;
+				MainLabel.ToolTip = this._info.Name;
+				MainLabel.Text = this._info.Name;
+				if (this._info.Warning.Length > 0) {
+					WarnIcon.ToolTip = this._info.Warning;
 					WarnIcon.Visibility = Visibility.Visible;
 					WarningColumn.Width = new GridLength(30);
 				}
 			} 
-		}
-
-		public string Label {
-			get {
-				return _label;
-			}
-			set {
-				this._label = value;
-				MainLabel.ToolTip = this._label;
-				MainLabel.Text = this._label;
-			}
-		}
-		public string Value {
-			get {
-				return MainEdit.Text; 
-			}
-			set {
-				MainEdit.ToolTip = value;
-				MainEdit.Text = value;
-			}
-		}
-		public bool IsLocked {
-			get {
-				return MainEdit.IsReadOnly;
-			} 
-			set {
-				MainEdit.IsReadOnly = value;
-			}
 		}
 
 		public ServiceInfo() {
@@ -69,8 +51,15 @@ namespace ZitiDesktopEdge {
         }
 
 		private void MainEdit_PreviewMouseUp(object sender, MouseButtonEventArgs e) {
-			var textbox = (sender as TextBox);
-			textbox.SelectAll();
+			(sender as TextBox).SelectAll();
+		}
+
+		private void WarnIcon_MouseUp(object sender, MouseButtonEventArgs e) {
+			OnMessage?.Invoke(this._info.Warning);
+		}
+
+		private void DetailIcon_MouseUp(object sender, MouseButtonEventArgs e) {
+			OnDetails?.Invoke(Info);
 		}
 	}
 }
