+++ DesktopEdge/Views/Screens/MFAScreen.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="ZitiDesktopEdge.RegistrationScreen"
+﻿<UserControl x:Class="Ziti.Desktop.Edge.Views.Screens.MFA"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:ZitiDesktopEdge"
+             xmlns:local="clr-namespace:Ziti.Desktop.Edge.Views.Screens"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
