+++ DesktopEdge/ZitiDesktopEdge.csproj
@@ -139,10 +139,13 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="ControllerScreen.xaml.cs">
-      <DependentUpon>ControllerScreen.xaml</DependentUpon>
+    <Compile Include="Views\ItemRenderers\MenuEditSearch.xaml.cs">
+      <DependentUpon>MenuEditSearch.xaml</DependentUpon>
     </Compile>
-    <Compile Include="IdentityDetails.xaml.cs">
+    <Compile Include="Views\ItemRenderers\MenuEditToggle.xaml.cs">
+      <DependentUpon>MenuEditToggle.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Views\Screens\IdentityDetails.xaml.cs">
       <DependentUpon>IdentityDetails.xaml</DependentUpon>
     </Compile>
     <Compile Include="MainMenu.xaml.cs">
@@ -156,9 +159,6 @@
     </Compile>
     <Compile Include="Views\ItemRenderers\MenuEditItem.xaml.cs">
       <DependentUpon>MenuEditItem.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="RegistrationScreen.xaml.cs">
-      <DependentUpon>RegistrationScreen.xaml</DependentUpon>
     </Compile>
     <Compile Include="Views\ItemRenderers\MenuIdentityItem.xaml.cs">
       <DependentUpon>MenuIdentityItem.xaml</DependentUpon>
@@ -177,12 +177,19 @@
     <Compile Include="Views\Controls\Toggler.xaml.cs">
       <DependentUpon>Toggler.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Views\Screens\MFAScreen.xaml.cs">
+      <DependentUpon>MFAScreen.xaml</DependentUpon>
+    </Compile>
     <Compile Include="WinAPI.cs" />
-    <Page Include="ControllerScreen.xaml">
-      <SubType>Designer</SubType>
+    <Page Include="Views\ItemRenderers\MenuEditSearch.xaml">
       <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
     </Page>
-    <Page Include="IdentityDetails.xaml">
+    <Page Include="Views\ItemRenderers\MenuEditToggle.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Views\Screens\IdentityDetails.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -217,10 +224,6 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="RegistrationScreen.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="Views\ItemRenderers\MenuIdentityItem.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -242,6 +245,10 @@
       <SubType>Designer</SubType>
     </Page>
     <Page Include="Views\Controls\Toggler.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Views\Screens\MFAScreen.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -375,7 +382,7 @@
     <Resource Include="Assets\Images\server.png" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Views\Screens\" />
+    <Resource Include="Assets\Images\detail.png" />
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Assets\Images\Warning.png" />
